+++ src/App.vue
@@ -6,8 +6,8 @@
 </template>
 
 <script setup lang="ts">
-import Navbar from "./components/NavBar.vue";
-import Main from "./components/Main.vue";
+import Navbar from "./components/Layout/NavBar.vue";
+import Main from "./components/Layout/Main.vue";
 import { defineComponent, onMounted } from "vue";
 import { toggleDarkMode } from "./utils";
