+++ src/components/Email/EmailForm.vue
@@ -100,7 +100,7 @@
 </template>
 
 <script setup lang="ts">
-import { useStore } from "../store";
+import { useStore } from "../../store";
 
 const store = useStore();
