+++ src/components/Email/EmailToggle.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script setup lang="ts">
-import { useStore } from "../store";
+import { useStore } from "../../store";
 
 const store = useStore();
 </script>
