+++ src/components/Layout/Main.vue
@@ -25,15 +25,15 @@
 
 <script setup lang="ts">
 import { defineComponent, onMounted } from "vue";
-import EmailToggle from "./EmailToggle.vue";
-import EmailForm from "./EmailForm.vue";
-import Introduction from "./Main/Introduction.vue";
-import WorkExperience from "./Main/WorkExperience.vue";
-import Projects from "./Main/Projects.vue";
-import Skills from "./Main/Skills.vue";
+import EmailToggle from "../Email/EmailToggle.vue";
+import EmailForm from "../Email/EmailForm.vue";
+import Introduction from "../Main/Introduction.vue";
+import WorkExperience from "../Main/WorkExperience.vue";
+import Projects from "../Main/Projects.vue";
+import Skills from "../Main/Skills.vue";
 import Footer from "./Footer.vue";
 
-import { useStore } from "../store";
+import { useStore } from "../../store";
 
 const store = useStore();
