+++ src/components/Skills/SkillItem.vue
@@ -1,0 +1,13 @@
+<template>
+  <li
+    class="rounded-3xl p-[2px] bg-gradient-to-tr from-rose-500 to-fuchsia-600 h-64 m-4 transition duration-300 ease-in-out hover:bg-gradient-to-r drop-shadow-xl"
+  >
+    <div
+      class="bg-gradient-to-tr from-neutral-600 to-stone-800 grid grid-cols-5 w-full h-full rounded-3xl overflow-hidden"
+    ></div>
+  </li>
+</template>
+
+<script setup lang="ts"></script>
+
+<style scoped></style>
